// src/app/apps/bizzcalc/models/Employee.ts
export interface Employee {
    Id?: number;
    employmentMonth?: string;
    jobTitle?: string;
    employmentReviewMonth?: string;
    employmentMonths?: number;
    employeeRate?: string;
    employmentGrowthRate?: string;
    employeeGrowthRate?: string;
    numberOfEmployees?: number;
    employeeSalary?: number;
    reviewMonth?: number;
    reviewPercentage?: number;
    pageNumber?: number;
    // new fields
    EmployeeName?: string;
    MonthEmployed?: string;
    Amount?: number;
    Qty?: number;
    AdjustmentPerc?: number;
    ReviewMonth?: number;
  }